feat!: Support ESM & refresh dependencies (#337)
* feat!: Convert to ESM and cleanup external deps

Dependencies updates:
* Update chokidar
* Remove fs-extra
* Replace inquirer/rxjs with @clack/prompts
* Replace ora with listr2
* Remove threads

The project is now in esm, and require the config to be `.cjs` or `.mjs` (breaking change). The migration is part of the CLI experience to ease the process.

* Update slash

* Update prettier

* Extract utils for readability

* Replace `case` with `text-case`

* Update eslint & co

* Update husky

* Ensure the typecheck is done

Vitest doesn’t typecheck our example file (out of the scope)

* Fix subtasks

* Uniform functions style

* Move prettier out of the flow

`prettier.resolveConfig` don’t resolve folder anymore, this make it very hard to deal with… Let’s move it out of the flow, leaving the user choosing how to format the generated files. As example this is done adding `&& pretty-quick` after the generation

* Remove note on zod limitation (supported in v4)

* Add a note about prettier

* Deal with explicit extension in import (`moduleResolution` = "node16" | "nodenext")

* Use package command (linter included)

* Try using node to invoke the script

* Regenerate config